Task 1 - Creates single function for each slice, could be inpried to be mor greneric - Save slices to image file

+++ native_language_task/src/volume_slicer/meta_data.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub struct MetaData {
-	pub n_dims: u32,
-	pub dim_size: (u32, u32, u32),
+	pub n_dims: usize,
+	pub dim_size: (usize, usize, usize),
 	pub element_spacing: (f32, f32, f32),
 }
